# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthtrix Expense backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}/finance-approve:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: financeApprove
      tags: [finance]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalActionBody" }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReport" }
  /reports/{id}/finance-reject:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: financeReject
      tags: [finance]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalActionBody" }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReport" }
  /reports/{id}/gl-preview:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getGlPreview
      tags: [finance]
      summary: Preview the QuickBooks JournalEntry that will be posted
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlPreview" }
  /reports/{id}/post-to-qbo:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: postToQuickbooks
      tags: [finance]
      summary: Post an Approved report to QuickBooks (stub)
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalActionBody" }
      responses:
        "200":
          description: Posted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostToQuickbooksResponse" }
  /reports/{id}/retry-qbo:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: retryQuickbooksPost
      tags: [finance]
      summary: Retry a Sync Error report posting
      responses:
        "200":
          description: Retried
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostToQuickbooksResponse" }
...
components:
  schemas:
...
    Role:
      type: string
      enum:
        - Employee
        - Manager Approver
        - Finance Approver
        - Accounting Admin
        - System Admin
    WorkflowStatus:
      type: string
      enum:
        - Draft
        - Submitted
        - Manager Review
        - Changes Requested
        - Manager Approved
        - Finance Review
        - Finance Approved
        - Posted to QuickBooks
        - Ready for Payroll Reimbursement
        - Paid Through Payroll
        - Reconciled
        - Rejected
        - Voided
        - Sync Error
    PaymentMethod:
      type: string
      enum: [Personal Card, Cash, Company Card]
...
    QboTag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: ["string", "null"] }
        active: { type: boolean }
      required: [id, name, active]
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
          minItems: 1
          uniqueItems: true
      required: [id, fullName, roles]
...
    ExpenseReport:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayCode: { type: string }
        title: { type: string }
        description: { type: string }
        employee: { $ref: "#/components/schemas/UserRef" }
        departmentId: { type: ["string", "null"], format: uuid }
        departmentName: { type: ["string", "null"] }
        policy: { type: string }
        periodStart: { type: ["string", "null"], format: date }
        periodEnd: { type: ["string", "null"], format: date }
        status: { $ref: "#/components/schemas/WorkflowStatus" }
        total: { type: string }
        submittedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lineItems:
          type: array
          items: { $ref: "#/components/schemas/LineItem" }
        receipts:
          type: array
          items: { $ref: "#/components/schemas/Receipt" }
        editedSinceLastApproval:
          type: boolean
          description: >-
            True when at least one field-level content edit has been
            recorded *after* the most recent approval action on this
            report (or after creation when no approvals exist yet, except
            for the initial draft state). Surfaces an "Edited since last
            approval" banner on the manager / finance review pages so
            reviewers can see a re-edit happened without forcing a
            status auto-reset.
        tags:
          type: array
          description: >-
            QBO tags currently applied to this report. Sent on the next
            JournalEntry post.
          items: { $ref: "#/components/schemas/QboTag" }
      required:
        - id
        - displayCode
        - title
        - description
        - employee
        - policy
        - status
        - total
        - createdAt
        - updatedAt
        - lineItems
        - receipts
        - editedSinceLastApproval
        - tags
...
    LineItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        reportId: { type: string, format: uuid }
        occurredOn: { type: string, format: date }
        merchant: { type: string }
        description: { type: string }
        category: { type: string }
        amount: { type: string }
        paymentMethod: { $ref: "#/components/schemas/PaymentMethod" }
        needsReview: { type: boolean }
        receiptCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - reportId
        - occurredOn
        - merchant
        - description
        - category
        - amount
        - paymentMethod
        - needsReview
        - receiptCount
        - createdAt
...
    Receipt:
      type: object
      properties:
        id: { type: string, format: uuid }
        reportId: { type: ["string", "null"], format: uuid }
        lineItemId: { type: ["string", "null"], format: uuid }
        objectPath: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        uploadedById: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - objectPath
        - filename
        - mimeType
        - sizeBytes
        - uploadedById
        - createdAt
...
    ApprovalActionBody:
      type: object
      properties:
        comment: { type: ["string", "null"] }
...
    GlPreviewLine:
      type: object
      properties:
        account: { type: string }
        accountId:
          type: ["string", "null"]
          description: >-
            Durable QBO Chart-of-Accounts Id, when the GL mapping has been
            linked to a real account. Posting payloads prefer this over the
            name because Intuit matches AccountRef by `value` (Id).
        category: { type: string }
        amount: { type: string }
      required: [account, category, amount]
    GlPreview:
      type: object
      properties:
        reportId: { type: string, format: uuid }
        displayCode: { type: string }
        journalDate: { type: string, format: date }
        memo: { type: string }
        debits:
          type: array
          items: { $ref: "#/components/schemas/GlPreviewLine" }
        credits:
          type: array
          items: { $ref: "#/components/schemas/GlPreviewLine" }
        totalDebits: { type: string }
        totalCredits: { type: string }
        currency: { type: string }
      required:
        - reportId
        - displayCode
        - journalDate
        - memo
        - debits
        - credits
        - totalDebits
        - totalCredits
        - currency
    PostToQuickbooksResponse:
      type: object
      properties:
        report: { $ref: "#/components/schemas/ExpenseReport" }
        journalId: { type: ["string", "null"] }
        status:
          type: string
          enum: [posted, error]
        errorMessage: { type: ["string", "null"] }
      required: [report, status]
...